♻️ Refactor: BinaryContentService 및 Repository의 delete 메서드 반환 타입을 void로 변경, findById 메서드에 Optional 적용
- JCF 및 File Repository에 existsById 메서드 추가 및 관련 로직 반영
- BinaryContent 존재 유무 검증 로직 추가
- RuntimeException 메시지 한국어 추가

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -22,7 +22,8 @@ public class BasicBinaryContentService implements BinaryContentService {
 
     @Override
     public BinaryContent findById(UUID id) {
-        return binaryContentRepository.findById(id);
+        return binaryContentRepository.findById(id).orElseThrow(
+                () -> new RuntimeException("바이너리 컨텐츠 입력값 잘못된 듯? "+ id));
     }
 
     @Override
@@ -31,7 +32,10 @@ public class BasicBinaryContentService implements BinaryContentService {
     }
 
     @Override
-    public BinaryContent delete(UUID id) {
-        return binaryContentRepository.delete(id);
+    public void delete(UUID id) {
+        if(!binaryContentRepository.existsById(id)){
+            throw new RuntimeException("BinaryContent not found. id: "+ id);
+        }
+        binaryContentRepository.delete(id);
     }
 }
